Comment delete

+++ DomesticAppliencesForum/src/main/java/com/telerikacademy/domesticappliencesforum/services/CommentServiceImpl.java
@@ -1,6 +1,8 @@
 package com.telerikacademy.domesticappliencesforum.services;
 
+import com.telerikacademy.domesticappliencesforum.exceptions.UnauthorizedOperationException;
 import com.telerikacademy.domesticappliencesforum.models.Comment;
+import com.telerikacademy.domesticappliencesforum.models.Post;
 import com.telerikacademy.domesticappliencesforum.models.User;
 import com.telerikacademy.domesticappliencesforum.repositories.CommentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -36,12 +38,20 @@ public class CommentServiceImpl implements CommentService {
     }
 
     @Override
-    public void modify(Comment comment) {
+    public void modify(Comment comment,User user) {
+        if (!(comment.getCreatedByUser().getUsername().equals(user.getUsername()))){
+            throw new UnauthorizedOperationException("You're not authorized for this operation");
+        }
         commentRepository.modify(comment);
     }
 
     @Override
-    public void delete(int id) {
+    public void delete(int id, User user) {
+        Comment comment = commentRepository.getCommentById(id);
+        if(!user.isAdmin() || !comment.getCreatedByUser().getUsername().equals(user.getUsername())){
+            throw new UnauthorizedOperationException("Only admins can delete");
+        }
+        //TODO Admina ни не работи както трябва!!!
         commentRepository.delete(id);
     }
